Add tests + fix build.gradle file
The tests, specifically in add command parser and edit command parser
did not test some of the new parameter types in existing commands.

To address this, tests with valid and invalid values for each parameter
type are added in both add command parser test and edit command parser
test.

In addition, the following minor changes are made:
- Add a new dependency line in build.gradle for javafx media
- Add new valid and invalid field values in parser util