Remove duplicate person handling in EditCommand
Currently, the EditCommand handles editing a person to a duplicate one.
It operates on the assumption that the name is the primary identifier of
a person, and two people with the same name are duplicates.
However, we have changed the person's primary identifier to be
matriculation number, as it is possible to have different people with
the same name, but matriculation number is unique.
Furthermore, the current implementation of EditCommand never modifies
the matriculation number. Therefore, we can guarantee that EditCommand
will never turn an existing person into a duplicate person, as
the person's matriculation number will remain the same as when
the person was added.
Hence, We can safely remove handling of duplicate person in EditCommand.