Update remove command for event
Currently the remove command for both volunteers and events does not
remove the deleted volunteers from all events' list of participants,
and the deleted event from all volunteers' list of events
they are involved in respectively.

This commit changes the behavior of the delete command to ensure
that all the aforementioned issue is resolved, and the command
behaves as intended.