Add new packages and reorganize classes into packages
In the commands, parser, and ui packages, there are many classes inside
and related classes aren't grouped together. It makes the code less
organized and harder to manage without separating them into packages.

Let's create packages to separate the kinds of commands for the parser
and commands packages, and separate the ui cards and panels.