Edit logic in EmergencyContactCommand
Logic allows for user to input emergency contact with only one of name
or number. There is also pre existing logic that displays emergency
contacts with empty fields as "No Emergency Contact" in the UI. This
means that the user will get a success message upon inputting an
incomplete emergency contact, yet no emergency contact will be shown.

New logic in EmergencyContactCommand is added to check that the
emergency contact to be added does not have empty fields. If it does,
an exception is thrown.

Allows user to have more clarity when inputting an incomplete
emergency contact, so they know that they made an error and can reinput
the contact properly.