Refactor JSON parser in ExportCommand
After the tags field in AddressBook.json was introduced, ExportCommand.readAndParseJson broke because its RegEx formulas did not account for array fields.

Since readAndParseJson broke, the JSON could not be parsed and its data could not be written to AddressBook.csv. As a result, invoking export format/csv in the CLI would create a blank AddressBook.csv.

Let's refactor the readAndParseJson function to account for tags while making the implementation more intuitive. The RegEx-first approach will be replaced by an implementation that uses the Jackson API to process JSON data.

Removing RegEx is our first step to reinforcing the robustness of the implementation. If we introduce more complex fields in the future, the RegEx patterns required to process AddressBook.json could become more complex and difficult to maintain. On the other hand, the Jackson API was designed to handle JSON files. In the context of this function, Jackson functions are easier to use, faster, and more flexible than a RegEx implementation.