Update EZSTATE storage
Updated EZSTATE storage architecture.

This patch adds 4 essential JSONAdapted classes to facilitate
the storage of the Appointment class and the 3 coupled
classes of Date, From, To.

The implementation is rather straightforward, using the same
structure as JsonAdaptedPerson and JsonAdapatedTag to
ensure that the deserialization process from JSON to
a usable Model class occurs as intended.

Additionally, the corresponding JsonAdaptedPersonTest and
JsonSerializableAddressBookTest have been updated, including
typicalPersonsAddressBook.json and
duplicatePersonAddressBook.json.