Add GUI changes to list of contacts
Contacts can't be seen if they've attended the CCA session for the
current date.

This makes it hard for users to expediently see which members are not
present for the CCA session at one glance. As such, they must
individually iew each and every member to note their attendance.

Let's add a visual GUI component in the list that indicates if that
particular contact has been marked as present for the current
session.