Refactor TypicalProperties and add getTypicalPropertyList method
The TypicalProperties class has been refactored to improve package
organization and functionality.

The class has been moved to the seedu.address.testutil.property
package to align with other property-related utility classes,
improving the structure and consistency of the codebase.
A new method getTypicalPropertyList() has been added, which
returns a PropertyList populated with typical properties. This
method simplifies the creation of PropertyList instances for
testing purposes by utilizing the existing getTypicalProperties()
method to add properties to the list.

These changes improve the organization and usability of the
TypicalProperties class, making it easier to manage typical property
data in test cases.