Fix a bug where clicking on menu bar interferes with arrow keys
There is a bug where if the user clicks on the menu bar, the
focus will be transferred to the menu bar component. This
interferes with the up and down arrow keys used for
command history.

Let's change the UI so that when users click on the menu
bar, the commandBox will be unfocused so that users
must click on the command box again to focus on it.