feat: Add support for emergency contact
This commit adds a new 'emergency contact' feature.
Emergency contact is implemented as a phone number only.
It may be added while using the `add` command,
or edited using the `edit` command with the same prefix `ec/`.

This first version implements the emergency contact field
as a phone number since it is the most critical component of a contact.