Add Member Role Detection
Attendance status shows up indiscriminately for each contact.

May not be necessary for contacts which are not members of
the CCA.

Let's leverage on the special Member role to display the
attendance status only for contacts that are members. This
includes both the list view, as well as the contact info
when using the iew command.