Add sort command to sort students by field(s)
The application does not support sorting functionality.

This feature is crucial for users when they want to view the
student listed in a particular order.

Add a sort command to allow users to sort the students by
field(s) in ascending or descending order.

The sort command will be able to sort the students by the
following fields:
- name
- phone
- address
- grade level

In the process of implementation, the ModelManager was modified
to return a sorted list of students instead of filtered list,
and filter to always be followed by the sort.

Let's add the sort command to the application.