Split ParserUtilTest
ParserUtilTest currently contains the tests for Event and Volunteer
parser utils. It should be split up into 2 different java files
for easier reference.

This commit has edited the tests for EventParserUtil only.

The tests do not pass in this current commit as it is written to
accommodate changes made in a PR that has yet to be merged (#109).