Create a list that keeps track of appointments
We are unable to show the schedule independent of the search results for
patients.

Having a separate list that handles the patients that have valid
appointments to display will allow for simultaneous usage of both lists.

Let's implement an ObservableList to keep track of changes made to the
appointment list, as well as a FilteredList that can handle any
filtering.

Add a Predicate to verify if a person's appointment should be
displayed given a particular date.