Bug fix: Make Date predicate check within a range
AppointmentContainsDatePredicate only checks if the given date is equal
to the start or end date.

This will result in errors as an appointment that lasts multiple days
will not be properly checked for.

Let's make the test method in AppointmentContainsDatePredicate consider
the range of dates between the start and end date of the appointment.