Add UnmatchCommand
Currently, there is a match command that associates the contact to a
job based on their respective ID.

However, there is no way to un match a contact from the job.

Creating a UnmatchCommand class enables the user to un match a
candidate from their job.

List of files added:
- UnmatchCommand.java - the main logic in un matching the contact and
the job
- UnmatchCommandParser.java - the parser which process user input
before sending it to UnmatchCommand.java

files edited:
- AddressBookParser.java - Added UnmatchCommandParser.java so it is
visible to the addressbook