Add better functionality to close claim and insurance plan manager
There was a bug where more than 1 claim could not be added. This is
because the insurance plan manager used the equals method to check
if the plan exists in the manager, thus, after the first check it
becomes impossible to "get" the manager as the created instance by
the factory is no longer the same as that inside the manager.

Let's change the getting process here to be a little more lax
to prevent such bugs.