Fix testing methods for Insurance and Claim Commands
The testing utilities do not support comparison of states of
InsurancePlansManager objects as we purposefully implemented it
as a mutable object and did not include this object to the client's
equals method.

This causes weird testing behaviours as the expected and intended
models used in assertCommandSuccess methods do not really check if
the two InsurancePlansManager objects are indeed the same.

Let's introduce a new testing utility to check for changes in
InsurancePlanManager objects instead of clients.

This should make the testing alot easier and accurate for commands
related to Insurances and Claims.