Add View Specific Student Command
There are only two type of UI layouts for person details: one excluding
task list and one which only shows tasks.

It is not possible to view a specific student's details all at once,
which is inconvenient if the tutor wants to view all details.

Let's introduce a command that allows viewing a specific student's
details, splitting the screen so that there is more space for details,
while still showing the list of students.

This allows more information to be shown at once, making efficient
use of the UI space.