Remove meetup filtering section from user guide
The section on filtering meetups using the filter command has been
removed from the user guide. This feature is no longer supported, and
its inclusion in the guide could lead to confusion for users.

Key changes:
* Removed the entire Filtering meetups : filter section.
* Updated the guide to ensure that the information provided reflects
the current functionality of the system.

This change improves the accuracy of the user guide, ensuring users
have clear and up-to-date information on available features.