Abstract Course code and Course Name to classes.
Currently, course code and course name in Course are stored as strings,
with their formats and validation fields and methods stored in Course. We
should create classes for course code and course name, similar to
Name and Email, used to store information for Person. This makes it
easier to parse commands for Course, as we do not have to manually
perform validation during parsing Course commands.