Bug fix: Make Appointment list sorted
The primary functionality of the schedule feature is to show the
user's schedule for a particular day. Hence, it makes sense for
the appointment list to be sorted.

Doing so will improve user experience as the results will be sorted,
making it easier for them to track their activities.

Let's use a SortedList to wrap around the filteredList to achieve
this functionality.