Improve ClientWindow
Though the client window allows user to view more details about the patient, it also made it compulsory
for users to enter certain input via mouse, specifically when exiting the client window.

I have hence created a key tracker, for when the client window is focused, to assign keyPress to the
variable keyEvent, which is then compared to the keycode of the ESCAPE key.

This way when user is in the client window, from the viewClient command, they could, by pressing the
exit key on their keyboard, exit the window returning back to the main window. Furthermore, upon
returning the input text bar would be auto-focused, hence effectively removing the need for any
mouse input.