Refine FindClientTypeCommand
Search algorithm for FindClientType is more neat.
For example, if a client has client type: "Insurance Plan"
doing "fc Insurance Plan" will now return the client. Doing "fc Insur
Plan" will also return the client.

Previously the search algorithm converts the client types of a person
into a stream and checks if all keywords match any of the prefix of the client types.
(eg. a client type of "Insurance Plan" is considered one word"

Now, the search will first combine the keyword if there is more than
1 keyword (eg. "Insurance Plan"). If there is a matching client type,
returns the client.

If not, it will do the usual algorithm by checking if all keywords match
any of the prefix of the client types. Difference is that now, it also
checks the individual word within the Client type itself. (eg. client
type of Insurance Plan will be split into Insurance and Plan
separately)

Add new and update test cases for FindClientType