Refactor income command
Previously, everything was done inside income command, which is not
really good, and it does not abide by the law of demeter as well

Let's
*Refactor the income command so it abides by law of demeter as well as
coding standards

Note: made changes to Messages, IncomeCommand, ModelManager, Student,
OwedAmount and PaidAmount.

For OwedAmount and PaidAmount, since I only added getter methods, did
not add tests for them.
For Student, since there were already tests for getOwedAmount and
getPaidAmount, felt there was no need to test for getOwedAmountValue
and getPaidAmountValue