Update UG Feature List
Feature list is not sorted, and parameters are not standardised.

The order of features is not intuitive and makes it more difficult for
users to navigate, and the inconsistencies in parameter terms used
makes it harder to understand.

Let's update the feature list and categorise them according to
function, and standardise parameter terms used.

Sorting the list according to type of features allows easier
navigation, and standardising parameters makes it more comprehensible.