Add Parameter Table to UG
There is no parameter table that lists all the possible parameters
used in commands.

It may be confusing for users to figure out what the parameter formats
are and their constraints.

Let's add a parameter table to consolidate all the possible parameters
used.

Users will be able to refer to the table and figure out what the
constraints and formats are for each parameter easily.