Fix bug in profile switching feature
Resolved the issue where a new profile was saved at inconsistent
stages in the program, causing the feature to be unusable.

Now, data files can be assumed to be created once a user executes
the switch command. The new file will not appear immediately but
will be created before the user closes the window.