Update name field
Previously, name field is case sensitive. This means that users are
able to add people with the same name using different case setting.

However, this is prone to errors by the user and having the user
remember 2 persons with the same name by a separate means such as
appending it with their surname would be more meaningful.

Thus, the name field is now case insensitive and inputting people
with the same name (ignoring case) is not allowed.