Modify feedback messages to user
The feedback message is long winded and provides irrelevant info
to the user.

This makes the app more difficult to use as the user has to read
through alot of information to understand if their command was
successful or not.

The following command feedback messages were updated in a consistent
manner:

1. AddClaimCommand
2. AddInsuranceCommand
3. CloseClaimCommand
4. DeleteClaimCommand
5. DeleteInsuranceCommand