Fix Assign and Unassign error message
There was an issue with inconsistent error messages for assign and
unassign raised in #132. This commit resolves the issue raised,
and additionally made the following changes:

* Error message "Index is not a non-zero unsigned integer" is
replaced with "Index has to be a positive integer" when the user
inputs a negative or 0 value as an index.
* Error message displayed after user inputs "unassign" or "assign"
command with missing fields has been changed to the "MESSAGE_USAGE"
strings in their respective classes.