Change how assignment fields are stored.
Grading status and grade were both stored in assignment.

This is redundant since an empty grade implies assignment has not
been graded. Let's change it so that we only store a grade. This
reduces the need for unnecessary checking.

Submission status and grade are reworked to ensure validity as well.