Update formatting, diagrams and add diagrams
The delete command's diagram does not reflect the confirmation prompt.
In addition, the sequence and activity diagrams for add and edit commands are also not created yet,
and we can talk about them more in the implementation section.
Formatting of the dg is also inconsistent.

Updating the delete command's sequence diagram helps with this.
New diagrams for add and edit commands have also been created, and added
under the implementation section.
The following changes have been made to ensure consistency:
- Used ## headers instead of ** in most places, except last section,
  which can be justified, as it is a separate section.
- Removed ** if attribute is enclosed in ``.